pyproject.toml
[tool.poetry]
name = "tna-flask-application"
version = "0.1.0"
description = ""
authors = ["Andrew Hosgood <andrew.hosgood@nationalarchives.gov.uk>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.13"
flask = "^3.1.0"
requests = "^2.31.0"
flask-caching = "^2.1.0"
tna-frontend-jinja = "0.33.0"
flask-talisman = "^1.1.0"
pyyaml = "^6.0.2"
flask-wtf = "^1.2.2"
deepmerge = "^2.0"
email-validator = "^2.2.0"
flask-sqlalchemy = "~3.1"
flask-migrate = "~4.1"
psycopg2-binary = "~2.9"
flask-session = "~0.8"
redis = "~6.4"
python-statemachine = "^2.5.0"
boto3 = "~1.40"
sentry-sdk = {extras = ["flask"], version = "^2.41.0"}
markupsafe = "^3.0.3"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
requests-mock = "^1.12.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
include = '\.pyi?$'

[tool.pytest.ini_options]
testpaths = [
    "test",
]
